# docs/swagger.yaml
definitions:
  problemdetail.ProblemDetail:
    properties:
      context: {}
      detail:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: List all Users
      operationId: ListUsers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: List all Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new User
      operationId: CreateUser
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Create a new User
      tags:
      - users
  /api/v1/users/{userID}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      operationId: DeleteUser
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get one User
      operationId: GetOneUser
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Get one User
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Update a User
      operationId: UpdateUser
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Update a User
      tags:
      - users
  /api/v1/users/invite/{email}:
    post:
      consumes:
      - application/json
      description: Invite a User by email
      operationId: InviteUser
      parameters:
      - description: User's email
        in: path
        name: email
        required: true
        type: string
      - description: User invite data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Invite a User by email
      tags:
      - users
  /api/v1/users/query:
    get:
      consumes:
      - application/json
      description: Search all Users
      operationId: SearchUsers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problemdetail.ProblemDetail'
      summary: Search all Users
      tags:
      - users
swagger: "2.0"